Correctly show the static app preview if we don't have enough thumbnails.
